refactor grpc folder
