+++ cmd/shortener/main.go
@@ -7,10 +7,11 @@ import (
 	_ "github.com/shulganew/shear.git/docs" // docs is generated by Swag CLI, you have to import it.
 	"github.com/shulganew/shear.git/internal/app"
 	"github.com/shulganew/shear.git/internal/config"
-	"github.com/shulganew/shear.git/internal/grpcs/servgrpc"
+	"github.com/shulganew/shear.git/internal/handler/grpcs/servgrpc"
+	"github.com/shulganew/shear.git/internal/handler/http/router"
+	"github.com/shulganew/shear.git/internal/handler/http/servrest"
 	"github.com/shulganew/shear.git/internal/service"
-	"github.com/shulganew/shear.git/internal/web/router"
-	"github.com/shulganew/shear.git/internal/web/servrest"
+
 	"github.com/shulganew/shear.git/pkg/shutdown"
 	"go.uber.org/zap"
 )
