+++ internal/handler/http/rest/batch.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"encoding/json"
