+++ internal/handler/http/rest/del.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"encoding/json"
