+++ internal/handler/http/rest/example_test.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"fmt"
