+++ internal/handler/http/rest/json_test.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"context"
