+++ internal/handler/http/rest/ping.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"database/sql"
