+++ internal/handler/http/rest/ping_test.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"context"
