+++ internal/handler/http/rest/stat.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"encoding/json"
