+++ internal/handler/http/rest/url.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"errors"
