+++ internal/handler/http/rest/usersurls.go
@@ -1,4 +1,4 @@
-package handlers
+package rest
 
 import (
 	"encoding/binary"
